auth for view

+++ app/Http/Controllers/QuestionController.php
@@ -75,11 +75,10 @@ class QuestionController extends Controller
     {
 
         $question = Question::findOrfail($id);
-        if (Gate::allows('update-question', $question)) {
-
-            return view('questions.edit', compact('question'));
+        if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access denies');
         }
-        abort(403, 'Access denies');
+        return view('questions.edit', compact('question'));
 
     }
 
@@ -94,6 +93,9 @@ class QuestionController extends Controller
     {
         //
         $question = Question::findOrfail($id);
+        if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access denies');
+        }
 
         $question->update($request->only('title', 'body'));
 
@@ -111,6 +113,9 @@ class QuestionController extends Controller
         //
         $question = Question::findOrfail($id);
         // dd($question);
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denies');
+        }
         $question->delete();
         return redirect()->route('allquestion')->with('success', 'Your question has been deleted');
